Update `EditorconfigParser`
Update the code to use a mutable imperative approach in order to avoid
excessive allocations.

Update the logic to take into account the recent changes to the
EditorConfig specs. Inline comments have been deprecated. Symbols
`#` and `;` are considered as a start of comment _only_ if preceded
by zero or more white spaces.